Extract trace
