+++ ethcore/Cargo.toml
@@ -67,6 +67,7 @@ state-account = { path = "state-account" }
 stats = { path = "../util/stats" }
 tempdir = { version = "0.3", optional = true }
 time-utils = { path = "../util/time-utils" }
+trace = { path = "trace" }
 trace-time = "0.1"
 triehash-ethereum = { version = "0.2",  path = "../util/triehash-ethereum" }
 unexpected = { path = "../util/unexpected" }
