+++ ethcore/src/lib.rs
@@ -102,6 +102,7 @@ extern crate serde;
 extern crate stats;
 extern crate state_account;
 extern crate time_utils;
+extern crate trace;
 extern crate triehash_ethereum as triehash;
 extern crate unexpected;
 extern crate using_queue;
@@ -166,7 +167,8 @@ pub mod snapshot;
 pub mod spec;
 pub mod state;
 pub mod state_db;
-pub mod trace;
+// TODO: need to reexport trace?
+//pub mod trace;
 pub mod transaction_ext;
 pub mod verification;
