+++ ethcore/trace/src/types/mod.rs
